Move operation_derivative
