+++ tests/test_qubit_device.py
@@ -28,7 +28,6 @@ from pennylane.circuit_graph import CircuitGraph
 from pennylane.variable import Variable
 from pennylane.wires import Wires
 from pennylane.tape.measure import state
-from pennylane._qubit_device import operation_derivative
 from pennylane.tape import QNode, qnode
 
 mock_qubit_device_paulis = ["PauliX", "PauliY", "PauliZ"]
@@ -1190,64 +1189,3 @@ class TestAdjointJacobianQNode:
         grad_fd = params1.grad, params2.grad
 
         assert np.allclose(grad_adjoint, grad_fd)
-
-
-class TestOperationDerivative:
-    """Tests for operation_derivative function"""
-
-    def test_no_generator_raise(self):
-        """Tests if the function raises a ValueError if the input operation has no generator"""
-        op = qml.Rot(0.1, 0.2, 0.3, wires=0)
-
-        with pytest.raises(ValueError, match="Operation Rot does not have a generator"):
-            operation_derivative(op)
-
-    def test_multiparam_raise(self):
-        """Test if the function raises a ValueError if the input operation is composed of multiple
-        parameters"""
-        class RotWithGen(qml.Rot):
-            generator = [np.zeros((2, 2)), 1]
-
-        op = RotWithGen(0.1, 0.2, 0.3, wires=0)
-
-        with pytest.raises(ValueError, match="Operation RotWithGen is not written in terms of"):
-            operation_derivative(op)
-
-    def test_rx(self):
-        """Test if the function correctly returns the derivative of RX"""
-        p = 0.3
-        op = qml.RX(p, wires=0)
-
-        derivative = operation_derivative(op)
-
-        expected_derivative = 0.5 * np.array([[-np.sin(p / 2), -1j * np.cos(p / 2)],[-1j * np.cos(p / 2), - np.sin(p / 2)]])
-
-        assert np.allclose(derivative, expected_derivative)
-
-        op.inv()
-        derivative_inv = operation_derivative(op)
-        expected_derivative_inv = 0.5 * np.array([[-np.sin(p / 2), 1j * np.cos(p / 2)],[1j * np.cos(p / 2), -np.sin(p / 2)]])
-
-        assert not np.allclose(derivative, derivative_inv)
-        assert np.allclose(derivative_inv, expected_derivative_inv)
-
-    def test_phase(self):
-        """Test if the function correctly returns the derivative of PhaseShift"""
-        p = 0.3
-        op = qml.PhaseShift(p, wires=0)
-
-        derivative = operation_derivative(op)
-        expected_derivative = np.array([[0, 0], [0, 1j * np.exp(1j * p)]])
-        assert np.allclose(derivative, expected_derivative)
-
-    def test_cry(self):
-        """Test if the function correctly returns the derivative of CRY"""
-        p = 0.3
-        op = qml.CRY(p, wires=[0, 1])
-
-        derivative = operation_derivative(op)
-        expected_derivative = 0.5 * np.array([[0, 0, 0, 0], [0, 0, 0, 0],
-                                        [0, 0, -np.sin(p / 2), - np.cos(p / 2)],
-                                        [0, 0, np.cos(p / 2), - np.sin(p / 2)],
-                                        ])
-        assert np.allclose(derivative, expected_derivative)
